Updating activity messages for accuracy.  Also, adding source field to activity.

+++ platform/api/StampedAPI.py
@@ -871,7 +871,13 @@ class StampedAPI(AStampedAPI):
     @API_CALL
     def removeLinkedAccount(self, authUserId, service_name):
         if service_name not in ['facebook', 'twitter', 'netflix']:
+            logs.warning('Attempted to remove invalid linked account: %s' % service_name)
             raise StampedIllegalActionError("Invalid linked account: %s" % service_name)
+
+        account = self.getAccount(authUserId)
+        if account.auth_service == service_name:
+            logs.warning('Attempted to remove linked account used for auth service')
+            raise StampedIllegalActionError('Cannot remove a third-party account used for authorization.')
 
         self._accountDB.removeLinkedAccount(authUserId, service_name)
         return True
@@ -1637,7 +1643,7 @@ class StampedAPI(AStampedAPI):
             entity  = self._entityDB.getEntity(stamp.entity.entity_id)
 
             if action in actions and authUserId != stamp.user.user_id:
-                self._addActionCompleteActivity(authUserId, action, stamp.stamp_id, stamp.user.user_id)
+                self._addActionCompleteActivity(authUserId, action, source, stamp.stamp_id, stamp.user.user_id)
 
         return True
 
@@ -4099,11 +4105,12 @@ class StampedAPI(AStampedAPI):
                                                               requireRecipient = True,
                                                               unique = True)
 
-    def _addActionCompleteActivity(self, userId, action_name, stampId, friendId, body=None):
+    def _addActionCompleteActivity(self, userId, action_name, source, stampId, friendId, body=None):
         objects = ActivityObjectIds()
         objects.user_ids        = [ friendId ]
         objects.stamp_ids       = [ stampId ]
         self._addActivity('action_%s' % action_name, userId, objects,
+                                                             source = source,
                                                              body = body,
                                                              group = True,
                                                              groupRange = timedelta(days=1),
@@ -4112,6 +4119,7 @@ class StampedAPI(AStampedAPI):
     def _addActivity(self, verb,
                            userId,
                            objects,
+                           source=None,
                            body=None,
                            recipientIds=[],
                            requireRecipient=False,
@@ -4137,6 +4145,7 @@ class StampedAPI(AStampedAPI):
         self._activityDB.addActivity(verb           = verb,
                                      subject        = userId,
                                      objects        = objects,
+                                     source         = source,
                                      body           = body,
                                      recipientIds   = recipientIds,
                                      benefit        = benefit,
@@ -4152,7 +4161,6 @@ class StampedAPI(AStampedAPI):
     @API_CALL
     def getActivity(self, authUserId, scope, limit=20, offset=0):
         activityData, final = self._activityCache.getFromCache(limit, offset, scope=scope, authUserId=authUserId)
-        #activityData, final = self._getActivityFromCache(authUserId, scope, offset, limit)
 
         # Append user objects
         userIds     = {}
